+++ Console/Assets/BlueConsole/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## [1.1.1] - 19.11.2023
+## [1.1.1] - 20.11.2023
  - Added input history recalling
  - Added a check for EventSystem in ConsoleController
+ - Improved commands parameters formatting
+ - Fixed "ArgumentException" from regex matching while typing in console input
 
 ## [1.1.0] - 19.11.2023
  - Added input support for old "InputSystem"